requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
paho-mqtt==2.1.0
ollama==0.5.1
openai-whisper==20231117
SpeechRecognition==3.10.0
pyaudio==0.2.11
pydub==0.25.1